[smscd2] for userchannel manage can use

+++ smsd2/command/Channel.py
@@ -21,7 +21,7 @@ class ChannelItemController(object):
     def __init__(self, context):
         self.c = context
         self.table = Table('ChannelItem', self.c.meta, autoload=True, autoload_with=self.c.db)
-        
+        self.channel_list_t = Table('ChannelList', self.c.meta, autoload=True, autoload_with=self.c.db)
     def add(self, **args):
         try:
             ins = self.table.insert().values(**args)
@@ -32,12 +32,74 @@ class ChannelItemController(object):
     
     def delete(self, uid):
         try:
+            if self.__channel_item_used(uid):
+                return False
             d = self.table.delete(self.table.c.uid==uid)
             self.c.db.execute(d)
             return True
         except:
             return False
     
+    def __channel_item_used(self, id):
+        try:
+            sel = select([self.channel_list_t], self.channel_list_t.c.cm1==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.cm2==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.cm3==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.cu1==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.cu2==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.cu3==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.ct1==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.ct2==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            sel = select([self.channel_list_t], self.channel_list_t.c.ct3==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            
+            return False
+        except:
+            print_exc()
+            return False
     def update(self, uid, **args):
         try:
             up = self.table.update().where(self.table.c.uid == uid).values(**args)
@@ -91,7 +153,7 @@ class ChannelListController(object):
     def __init__(self, context):
         self.c = context
         self.table = Table('ChannelList', self.c.meta, autoload=True, autoload_with=self.c.db)
-    
+        self.user_t = Table('user', self.c.meta, autoload=True, autoload_with=self.c.db)
     def add(self, **args):
         try:
             ins = self.table.insert().values(**args)
@@ -102,12 +164,26 @@ class ChannelListController(object):
     
     def delete(self, uid):
         try:
+            if self.__channel_list_used(uid):
+                return False
             d = self.table.delete(self.table.c.uid==uid)
             self.c.db.execute(d)
             return True
         except:
             return False
     
+    def __channel_list_used(self, id):
+        try:
+            sel = select([self.user_t], self.user_t.c.channel_list_id==id)
+            res = self.c.db.execute(sel)
+            r =  res.fetchone()
+            if r:
+                return True
+            else:
+                return False
+        except:
+            print_exc()
+            return False
     def update(self, uid, **args):
         try:
             up = self.table.update().where(self.table.c.uid == uid).values(**args)
